hacking json contracts
