+++ src/Tochka.JsonRpc.Common/Models/Response/Response.cs
@@ -12,7 +12,6 @@ namespace Tochka.JsonRpc.Common.Models.Response
         /// <summary>
         /// 
         /// </summary>
-        /// <example>2.0</example>
         public string Jsonrpc { get; set; } = JsonRpcConstants.Version;
 
         public TResult Result { get; set; }
