+++ src/WebApplication1/Controllers/TestData.cs
@@ -1,0 +1,8 @@
+﻿namespace WebApplication1.Controllers
+{
+    public class TestData
+    {
+        public int Value { get; set; }
+        public string AnotherValue { get; set; }
+    }
+}
