+++ src/WebApplication1/Controllers/TestData2.cs
@@ -1,0 +1,12 @@
+﻿using System.Text.Json.Serialization;
+using Newtonsoft.Json;
+
+namespace WebApplication1.Controllers
+{
+    public class TestData2
+    {
+        [JsonProperty("testdata2_value_JsonProperty")]
+        public int Value { get; set; }
+        public string AnotherValue { get; set; }
+    }
+}
