+++ src/WebApplication1/WebApplication1.xml
@@ -4,41 +4,45 @@
         <name>WebApplication1</name>
     </assembly>
     <members>
-        <member name="M:WebApplication1.Controllers.KostylProvider.PatchPath(Microsoft.AspNetCore.Mvc.ApiExplorer.ApiDescription,System.String)">
-            <summary>
-            Add #method to JsonRpc urls to make them different for swagger
-            </summary>
-            <param name="apiDescription"></param>
-            <param name="methodMetadata"></param>
-        </member>
-        <member name="M:WebApplication1.Controllers.KostylProvider.PatchParams(Microsoft.AspNetCore.Mvc.ApiExplorer.ApiDescription,Tochka.JsonRpc.Server.Models.MethodMetadata,System.String)">
-            <summary>
-            Hide separate JsonRpc-bound parameters and wrap them into Request`T. T is emitted dynamically to allow swagger schema generation
-            </summary>
-            <param name="apiDescription"></param>
-            <param name="methodMetadata"></param>
-            <param name="methodName"></param>
-        </member>
-        <member name="M:WebApplication1.Controllers.KostylProvider.PatchRequestFormats(Microsoft.AspNetCore.Mvc.ApiExplorer.ApiDescription)">
-            <summary>
-            Set request content-type as application/json
-            </summary>
-            <param name="apiDescription"></param>
-        </member>
-        <member name="M:WebApplication1.Controllers.KostylProvider.PatchResponseTypes(Microsoft.AspNetCore.Mvc.ApiExplorer.ApiDescription)">
-            <summary>
-            Wrap action response type into Response`T, set content-type as application/json and HTTP code 200
-            </summary>
-            <param name="apiDescription"></param>
-        </member>
-        <member name="P:WebApplication1.Controllers.KostylProvider.Order">
-            <summary>
-            Arbitrary number greater than -1000 in DefaultApiDescriptionProvider
-            </summary>
-        </member>
         <member name="T:WebApplication1.Controllers.ModelTypeEmitter">
             <summary>
             Creates types for proper schema generation. See https://jonathancrozier.com/blog/time-to-reflect-how-to-add-properties-to-a-c-sharp-object-dynamically-at-runtime
+            </summary>
+        </member>
+        <member name="M:WebApplication1.Controllers.WtfProvider.GetUniquePath(System.String,System.String)">
+            <summary>
+            Add #method to JsonRpc urls to make them different for swagger
+            </summary>
+        </member>
+        <member name="P:WebApplication1.Controllers.WtfProvider.JsonRequestFormat">
+            <summary>
+            Request content-type application/json
+            </summary>
+        </member>
+        <member name="M:WebApplication1.Controllers.WtfProvider.WrapResponseType(System.Type)">
+            <summary>
+            Wrap action response type into Response`T, set content-type application/json and HTTP code 200
+            </summary>
+        </member>
+        <member name="M:WebApplication1.Controllers.WtfProvider.GetParameterDescriptions(Microsoft.AspNetCore.Mvc.ApiExplorer.ApiDescription,Tochka.JsonRpc.Server.Models.MethodMetadata)">
+            <summary>
+            Wrap JsonRpc-bound parameters into Request`T. T is compiled dynamically to allow swagger schema generation
+            </summary>
+            <param name="defaultDescription"></param>
+            <param name="methodMetadata"></param>
+            <returns></returns>
+        </member>
+        <member name="M:WebApplication1.Controllers.WtfProvider.GetRequestType(Tochka.JsonRpc.Server.Models.MethodMetadata)">
+            <summary>
+            Generate type which describes whole request object with all parameters bound by json rpc
+            </summary>
+            <param name="defaultDescription"></param>
+            <param name="methodMetadata"></param>
+            <returns></returns>
+        </member>
+        <member name="P:WebApplication1.Controllers.WtfProvider.Order">
+            <summary>
+            Arbitrary number greater than -1000 in DefaultApiDescriptionProvider
             </summary>
         </member>
     </members>
